Do not wait for test queue tasks to complete after workers completes
pytest-parallel waits for all tasks that have been picked up by workers
to be marked as completed. However, workers may have been killed (e.g.,
in the case of a timeout), and thus failed to mark the tasks as
completed. In this case, the main thread would wait indefinitely.

Since there is no apparent need to wait for the task completion (since
that is implied by all workers completing), we remove it, fixing the
hang issue in the case of a failed worker.