Unregister terminalreporter in workers
The forked workers run pytest_runtest_protocol, which in turn runs
the following hooks:
- pytest_runtest_logstart
- pytest_runtest_logreport
- pytest_runtest_logfinish
The built-in "terminalreporter" plugin hooks to all of the above to print
the progress of the test suite (printing each completed tests and other
info such as the test-suite completion percentage). This results in
interleaved output, as well as duplicating the output printed by the
main thread which is also running pytest_runtest_logreport.

We thus unregister terminalreporter in the workers before running any
test, leaving main in charge of printing test results.

We also reinstate the normal behaviour of terminalreporter that was
modified in ParallelRunner by setting "reporter.showfspath = False" and
"reporter._show_progress_info = False" in an incorrect attempt to
prevent output mangling.