Override workers when number of tests to execute is less
A number of pytest sessions are hanging at the end and it is not 100%
clear what the cause is. However, some experimentation with the number
of workers and tests to execute has shown:

  - When they exceed the number of tests to execute, the hang at the
    end of the pytest session is more likely to occur
  - When they are equal to or less than the tests to execute, the hang
    has yet to occur

Therefore, this speculative fix overrides the `workers` to be the
number of tests to execute when it exceeds the originally configured
number.